third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/the-form-element/form-autocomplete-expected.txt

This is a testharness.js-based test.
Found 59 tests; 0 PASS, 59 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL form autocomplete attribute missing assert_equals: expected "on" but got ""
FAIL form autocomplete attribute on assert_equals: expected "on" but got ""
FAIL form autocomplete attribute off assert_equals: expected "off" but got ""
FAIL form autocomplete attribute invalid assert_equals: expected "on" but got ""
FAIL on is an allowed autocomplete field name assert_equals: expected "on" but got " ON\t"
FAIL off is an allowed autocomplete field name assert_equals: expected "off" but got " OFF\t"
FAIL name is an allowed autocomplete field name assert_equals: expected "name" but got " NAME\t"
FAIL honorific-prefix is an allowed autocomplete field name assert_equals: expected "honorific-prefix" but got " HONORIFIC-PREFIX\t"
FAIL given-name is an allowed autocomplete field name assert_equals: expected "given-name" but got " GIVEN-NAME\t"
FAIL additional-name is an allowed autocomplete field name assert_equals: expected "additional-name" but got " ADDITIONAL-NAME\t"
FAIL family-name is an allowed autocomplete field name assert_equals: expected "family-name" but got " FAMILY-NAME\t"
FAIL honorific-suffix is an allowed autocomplete field name assert_equals: expected "honorific-suffix" but got " HONORIFIC-SUFFIX\t"
FAIL nickname is an allowed autocomplete field name assert_equals: expected "nickname" but got " NICKNAME\t"
FAIL username is an allowed autocomplete field name assert_equals: expected "username" but got " USERNAME\t"
FAIL new-password is an allowed autocomplete field name assert_equals: expected "new-password" but got " NEW-PASSWORD\t"
FAIL current-password is an allowed autocomplete field name assert_equals: expected "current-password" but got " CURRENT-PASSWORD\t"
FAIL organization-title is an allowed autocomplete field name assert_equals: expected "organization-title" but got " ORGANIZATION-TITLE\t"
FAIL organization is an allowed autocomplete field name assert_equals: expected "organization" but got " ORGANIZATION\t"
FAIL street-address is an allowed autocomplete field name assert_equals: expected "street-address" but got " STREET-ADDRESS\t"
FAIL address-line1 is an allowed autocomplete field name assert_equals: expected "address-line1" but got " ADDRESS-LINE1\t"
FAIL address-line2 is an allowed autocomplete field name assert_equals: expected "address-line2" but got " ADDRESS-LINE2\t"
FAIL address-line3 is an allowed autocomplete field name assert_equals: expected "address-line3" but got " ADDRESS-LINE3\t"
FAIL address-level4 is an allowed autocomplete field name assert_equals: expected "address-level4" but got " ADDRESS-LEVEL4\t"
FAIL address-level3 is an allowed autocomplete field name assert_equals: expected "address-level3" but got " ADDRESS-LEVEL3\t"
FAIL address-level2 is an allowed autocomplete field name assert_equals: expected "address-level2" but got " ADDRESS-LEVEL2\t"
FAIL address-level1 is an allowed autocomplete field name assert_equals: expected "address-level1" but got " ADDRESS-LEVEL1\t"
FAIL country is an allowed autocomplete field name assert_equals: expected "country" but got " COUNTRY\t"
FAIL country-name is an allowed autocomplete field name assert_equals: expected "country-name" but got " COUNTRY-NAME\t"
FAIL postal-code is an allowed autocomplete field name assert_equals: expected "postal-code" but got " POSTAL-CODE\t"
FAIL cc-name is an allowed autocomplete field name assert_equals: expected "cc-name" but got " CC-NAME\t"
FAIL cc-given-name is an allowed autocomplete field name assert_equals: expected "cc-given-name" but got " CC-GIVEN-NAME\t"
FAIL cc-additional-name is an allowed autocomplete field name assert_equals: expected "cc-additional-name" but got " CC-ADDITIONAL-NAME\t"
FAIL cc-family-name is an allowed autocomplete field name assert_equals: expected "cc-family-name" but got " CC-FAMILY-NAME\t"
FAIL cc-number is an allowed autocomplete field name assert_equals: expected "cc-number" but got " CC-NUMBER\t"
FAIL cc-exp is an allowed autocomplete field name assert_equals: expected "cc-exp" but got " CC-EXP\t"
FAIL cc-exp-month is an allowed autocomplete field name assert_equals: expected "cc-exp-month" but got " CC-EXP-MONTH\t"
FAIL cc-exp-year is an allowed autocomplete field name assert_equals: expected "cc-exp-year" but got " CC-EXP-YEAR\t"
FAIL cc-csc is an allowed autocomplete field name assert_equals: expected "cc-csc" but got " CC-CSC\t"
FAIL cc-type is an allowed autocomplete field name assert_equals: expected "cc-type" but got " CC-TYPE\t"
FAIL transaction-currency is an allowed autocomplete field name assert_equals: expected "transaction-currency" but got " TRANSACTION-CURRENCY\t"
FAIL transaction-amount is an allowed autocomplete field name assert_equals: expected "transaction-amount" but got " TRANSACTION-AMOUNT\t"
FAIL language is an allowed autocomplete field name assert_equals: expected "language" but got " LANGUAGE\t"
FAIL bday is an allowed autocomplete field name assert_equals: expected "bday" but got " BDAY\t"
FAIL bday-day is an allowed autocomplete field name assert_equals: expected "bday-day" but got " BDAY-DAY\t"
FAIL bday-month is an allowed autocomplete field name assert_equals: expected "bday-month" but got " BDAY-MONTH\t"
FAIL bday-year is an allowed autocomplete field name assert_equals: expected "bday-year" but got " BDAY-YEAR\t"
FAIL sex is an allowed autocomplete field name assert_equals: expected "sex" but got " SEX\t"
FAIL url is an allowed autocomplete field name assert_equals: expected "url" but got " URL\t"
FAIL photo is an allowed autocomplete field name assert_equals: expected "photo" but got " PHOTO\t"
FAIL tel is an allowed autocomplete field name assert_equals: expected "tel" but got " TEL\t"
FAIL tel-country-code is an allowed autocomplete field name assert_equals: expected "tel-country-code" but got " TEL-COUNTRY-CODE\t"
FAIL tel-national is an allowed autocomplete field name assert_equals: expected "tel-national" but got " TEL-NATIONAL\t"
FAIL tel-area-code is an allowed autocomplete field name assert_equals: expected "tel-area-code" but got " TEL-AREA-CODE\t"
FAIL tel-local is an allowed autocomplete field name assert_equals: expected "tel-local" but got " TEL-LOCAL\t"
FAIL tel-local-prefix is an allowed autocomplete field name assert_equals: expected "tel-local-prefix" but got " TEL-LOCAL-PREFIX\t"
FAIL tel-local-suffix is an allowed autocomplete field name assert_equals: expected "tel-local-suffix" but got " TEL-LOCAL-SUFFIX\t"
FAIL tel-extension is an allowed autocomplete field name assert_equals: expected "tel-extension" but got " TEL-EXTENSION\t"
FAIL email is an allowed autocomplete field name assert_equals: expected "email" but got " EMAIL\t"
FAIL impp is an allowed autocomplete field name assert_equals: expected "impp" but got " IMPP\t"
Harness: the test ran to completion.
